refactor(gamma-platform): share the policy catalog service between features
Hoisted out of shared-policy-groups with the protocol as a parameter, and typed documentation-ext as the document it returns.
